app/requirements.txt
asgiref==3.8.1
colorama==0.4.6
diff-match-patch==20241021
Django==5.1.7
django-browser-reload==1.18.0
django-countries==7.6.1
django-import-export==4.3.7
django-tailwind==3.8.0
django-widget-tweaks==1.5.0
pillow==11.1.0
qrcode==8.0
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.12.2
tzdata==2025.1
gunicorn
